Rework license header checks
Now three version for the license header checks are provided. One to
check all files in the repository on HEAD, one to check all files
changed on a branch with respect to master, and finally to check all
indexed files prior to a commit. The last version may serve as script
for pre-commit hooks soon.